Changed the names for the UserDAO's and fixed the Func layer
