+++ src/main/java/data/memory/MemoryUserDAO.java
@@ -7,7 +7,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 
-public class UserDAO implements IUserDAO {
+public class MemoryUserDAO implements IUserDAO {
 
     HashMap<Integer,UserDTO> userHMap = new HashMap<Integer, UserDTO>();
 
@@ -16,7 +16,6 @@ public class UserDAO implements IUserDAO {
         return userHMap.get(userID);
     }
 
-
     public void updateUser(UserDTO user) {
         UserDTO tempUser = user;
         userHMap.remove(user.getId());
@@ -24,7 +23,6 @@ public class UserDAO implements IUserDAO {
         userHMap.put(tempUser.getId(),tempUser);
     }
 
-    //TODO snak igennem alle de her parametre fordi nogle af dem giver ingen megning.
     @Override
     public void createUser(UserDTO user) {
       userHMap.put(user.getId(),user);
